* Implemented the PlayerProgressTracker that saves the users current progress in the stream every 8 seconds * When the stream ends the progress is reset * When the stream is started, the progress is resumed if there is any saved * When a new URL is chosen, the progress is reset

+++ lib/src/test/java/com/memtrip/exoeasy/broadcast/AudioStateBroadcastReceiverTest.kt
@@ -123,11 +123,11 @@ class AudioStateBroadcastReceiverTest : Spek({
                 }.thenReturn(10)
 
                 on {
-                    getIntExtra("EXTRA_BROADCAST_PROGRESS_POSITION", 0)
+                    getLongExtra("EXTRA_BROADCAST_PROGRESS_POSITION", 0)
                 }.thenReturn(50)
 
                 on {
-                    getIntExtra("EXTRA_BROADCAST_PROGRESS_DURATION", 0)
+                    getLongExtra("EXTRA_BROADCAST_PROGRESS_DURATION", 0)
                 }.thenReturn(100)
             }
